Internal: Merged ResizerTopBound class back into ResizeContext.
